Changed the argument type on internal_dispose_GC from int handle to long handle.
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@623 bb130017-240f-0410-9613-9c2c827820fb

+++ core/cc.warlock.rcp/src/cc/warlock/rcp/ui/WarlockText.java
@@ -572,7 +572,7 @@ public class WarlockText implements LineBackgroundListener, Drawable {
 		return textWidget.getContent();
 	}
 
-	public void internal_dispose_GC(int handle, GCData data) {
+	public void internal_dispose_GC(long handle, GCData data) {
 		textWidget.internal_dispose_GC(handle, data);
 	}
